+++ kernel/tasking/Syscalls.h
@@ -4,4 +4,4 @@
 
 #include <libsystem/Common.h>
 
-int task_do_syscall(Syscall syscall, int arg0, int arg1, int arg2, int arg3, int arg4);
+uintptr_t task_do_syscall(Syscall syscall, uintptr_t arg0, uintptr_t arg1, uintptr_t arg2, uintptr_t arg3, uintptr_t arg4);
